move handling archival objects to its own file
