+++ app/actions/aspaceSearch.ts
@@ -1,3 +1,7 @@
+/*
+The entry point for all aspaceSearch functions is the searchByUrl function.
+It is called by: page.tsx > RecordSearchForm > searchCatalogByAny@actions/catalogSearch
+*/
 'use server';
 import {
   AspaceClient,
@@ -19,12 +23,17 @@ import type {
   CatalogSearchResult,
 } from '@/lib/catalogs/types';
 import { ZodError } from 'zod';
-import { HandleMissingInstances } from './aspaceHandleMissingInstances';
+import { HandleMissingInstances } from './aspace/handleMissingInstances';
+import { getArchivalObjectData } from './aspace/handleArchivalObjects';
 
 export interface ArchivalObjectExtraInfo {
   numItems: number;
+  sumSeries: number;
+  numSubseries: number;
+  summaryInfo: string;
   firstItemUrl: string;
   firstRecordArgusData: { bibData: BibDataDraft; itemData: ItemDataDraft };
+  items: any[];
 }
 
 export async function getClient() {
@@ -110,37 +119,7 @@ export async function searchByUrl(url: string, client: AspaceClient) {
       // https://archivesstaff.lib.miamioh.edu/api/repositories/2/archival_objects/13282
       // https://archivesstaff.lib.miamioh.edu/api/repositories/2/archival_objects/5616
       case /repositories\/\d+\/archival_objects\/\d+/.test(url): {
-        logger.verbose('aspaceSearch.searchByUrl found archivalObjects');
-
-        const parsed = repoArchivalObjectSchema.parse(raw);
-
-        /* 
-          If the archival object doesn't have any instances, try finding some 
-          by searching down the tree of the main resource
-        */
-        if (parsed.instances.length == 0) {
-          const parentResourceUrl = parsed.resource.ref;
-          logger.verbose(`parentResourceUrl: ${parentResourceUrl}`);
-          const originalUri = url.match(/\/repositories\/.*/);
-          logger.verbose(`originalUri: ${originalUri}`);
-          if (originalUri !== null) {
-            extraInfo = await HandleMissingInstances(
-              originalUri.toString(),
-              parentResourceUrl,
-              client,
-            );
-          }
-        }
-
-        let argusData;
-        if (Object.keys(extraInfo).length > 0) {
-          argusData = repoArchivalObjectToDraft(parsed, publicUrl, extraInfo);
-        } else {
-          argusData = repoArchivalObjectToDraft(parsed, publicUrl);
-        }
-        logger.verbose(argusData);
-        logger.verbose('type: archival objects');
-        return argusData;
+        return await getArchivalObjectData(raw, url, publicUrl, client);
         break;
       }
 
@@ -251,82 +230,83 @@ function repoTopContainerToDraft(data: RepoTopContainer, url: string) {
  * repoArchivalObjectToDraft
  */
 
-const getItems = (data: RepoArchivalObject) => {
-  if (data.instances.length > 0) {
-    return data.instances.map((item, index) => ({
-      clientKey: `item-${index}`,
-      barcode: '',
-      box: '',
-      call_number:
-        callNumberOverrides.archivalObject?.item?.(item, data) ??
-        item.sub_container?.top_container?._resolved?.display_string ??
-        '',
-      copy_id: '',
-      description:
-        item.sub_container?.top_container?._resolved?.long_display_string ?? '',
-      folder: '',
-      location_code: data.repository._resolved?.slug ?? '',
-      location_name: data.repository._resolved?.name ?? '',
-      ms: '',
-    }));
-  } else {
-    if (callNumberOverrides.archivalObject?.allItems) {
-      const callNumbers = callNumberOverrides.archivalObject?.allItems(data);
-      return callNumbers.map((callNumber, index) => ({
-        clientKey: `item-${index}`,
-        barcode: '',
-        box: '',
-        call_number: callNumber ?? '',
-        copy_id: '',
-        description: '',
-        folder: '',
-        location_code: data.repository._resolved?.slug ?? '',
-        location_name: data.repository._resolved?.name ?? '',
-        ms: '',
-      }));
-    }
-  }
-};
+// const getItems = (data: RepoArchivalObject) => {
+//   if (data.instances.length > 0) {
+//     return data.instances.map((item, index) => ({
+//       clientKey: `item-${index}`,
+//       barcode: '',
+//       box: '',
+//       call_number:
+//         callNumberOverrides.archivalObject?.item?.(item, data) ??
+//         item.sub_container?.top_container?._resolved?.display_string ??
+//         '',
+//       copy_id: '',
+//       description:
+//         item.sub_container?.top_container?._resolved?.long_display_string ?? '',
+//       folder: '',
+//       location_code: data.repository._resolved?.slug ?? '',
+//       location_name: data.repository._resolved?.name ?? '',
+//       ms: '',
+//     }));
+//   } else {
+//     if (callNumberOverrides.archivalObject?.allItems) {
+//       const callNumbers = callNumberOverrides.archivalObject?.allItems(data);
+//       return callNumbers.map((callNumber, index) => ({
+//         clientKey: `item-${index}`,
+//         barcode: '',
+//         box: '',
+//         call_number: callNumber ?? '',
+//         copy_id: '',
+//         description: '',
+//         folder: '',
+//         location_code: data.repository._resolved?.slug ?? '',
+//         location_name: data.repository._resolved?.name ?? '',
+//         ms: '',
+//       }));
+//     }
+//   }
+// };
 
-function repoArchivalObjectToDraft(
-  data: RepoArchivalObject,
-  url: string,
-  extraInfo?: any,
-) {
-  const bibData: BibDataDraft = {
-    author:
-      data.linked_agents
-        .filter((entry) => entry.role == 'creator')
-        .map((entry) => entry._resolved?.names[0].sort_name)
-        .join('; ') ?? 'Unknown',
-    callNumber:
-      callNumberOverrides.archivalObject?.bib?.(data, extraInfo) ??
-      data.instances
-        .map(
-          (instance) =>
-            instance.sub_container.top_container._resolved?.indicator,
-        )
-        .join('; ') ??
-      '',
-    itemTitle: titleOverrides.archivalObject?.(data) ?? data.title,
-    catalog: 'ASPACE',
-    catalogId: data.uri,
-    catalogIdType: 'uri',
-    location_codes: data.repository._resolved?.slug ?? '',
-    location_display: data.repository._resolved?.name ?? '',
-    notes: '',
-    pub_date:
-      (data.dates &&
-        data.dates[0] &&
-        (data.dates[0].begin || data.dates[0].end) &&
-        `${data.dates[0]?.begin ?? ''} - ${data.dates[0]?.end ?? ''}`) ??
-      '',
-    publisher: null,
-    totalItems: 1,
-    url: url,
-  };
+// function repoArchivalObjectToDraft(
+//   data: RepoArchivalObject,
+//   url: string,
+//   extraInfo?: any,
+// ) {
+//   console.log(`EXTRA INFO: ${JSON.stringify(extraInfo)}`);
+//   const bibData: BibDataDraft = {
+//     author:
+//       data.linked_agents
+//         .filter((entry) => entry.role == 'creator')
+//         .map((entry) => entry._resolved?.names[0].sort_name)
+//         .join('; ') ?? 'Unknown',
+//     callNumber:
+//       callNumberOverrides.archivalObject?.bib?.(data, extraInfo) ??
+//       data.instances
+//         .map(
+//           (instance) =>
+//             instance.sub_container.top_container._resolved?.indicator,
+//         )
+//         .join('; ') ??
+//       '',
+//     itemTitle: titleOverrides.archivalObject?.(data) ?? data.title,
+//     catalog: 'ASPACE',
+//     catalogId: data.uri,
+//     catalogIdType: 'uri',
+//     location_codes: data.repository._resolved?.slug ?? '',
+//     location_display: data.repository._resolved?.name ?? '',
+//     notes: '',
+//     pub_date:
+//       (data.dates &&
+//         data.dates[0] &&
+//         (data.dates[0].begin || data.dates[0].end) &&
+//         `${data.dates[0]?.begin ?? ''} - ${data.dates[0]?.end ?? ''}`) ??
+//       '',
+//     publisher: null,
+//     totalItems: 1,
+//     url: url,
+//   };
 
-  const itemData: ItemDataDraft[] = getItems(data) ?? [];
+//   const itemData: ItemDataDraft[] = getItems(data) ?? [];
 
-  return { bibData, itemData };
-}
+//   return { bibData, itemData };
+// }
